Companion for #Substrate-8955

+++ runtime/darwinia/src/pallets/system.rs
@@ -14,7 +14,7 @@ impl Filter<Call> for BaseFilter {
 
 frame_support::parameter_types! {
 	pub const Version: RuntimeVersion = VERSION;
-	pub const SS58Prefix: u8 = 18;
+	pub const SS58Prefix: u16 = 18;
 }
 
 impl Config for Runtime {
